--- custom_ops/gpu_ops/noauxtc_kernel_dev.h
@@ -1,1173 +1,0 @@
-// Copyright (c) 2025 PaddlePaddle Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// This code is partially inspired by and references the implementation found
-// in NVIDIA TRTLLM.
-#pragma once
-#include <cooperative_groups.h>
-#include <cooperative_groups/reduce.h>
-#include <cuda/std/limits>
-#include "helper.h"
-
-namespace cg = cooperative_groups;
-
-constexpr unsigned FULL_WARP_MASK = 0xffffffff;
-constexpr int32_t BLOCK_SIZE = 512;
-constexpr int32_t NUM_WARPS_PER_BLOCK = BLOCK_SIZE / WARP_SIZE;
-
-template <typename T_OUT, typename T_IN>
-__device__ inline T_OUT cuda_cast(T_IN val) {
-  return val;
-}
-
-template <>
-__device__ inline float cuda_cast<float, __nv_bfloat16>(__nv_bfloat16 val) {
-  return __bfloat162float(val);
-}
-
-template <typename T>
-__device__ inline T neg_inf() {
-  // cuda::std::numeric_limits<T>::infinity() returns `0` for [T=bf16 or fp16]
-  // so we need to cast from fp32
-  return cuda_cast<T, float>(-cuda::std::numeric_limits<float>::infinity());
-}
-
-namespace warp_topk {
-
-template <int size, typename T>
-__host__ __device__ constexpr T round_up_to_multiple_of(T len) {
-  if (len == 0) {
-    return 0;
-  }
-  return ((len - 1) / size + 1) * size;
-}
-
-template <typename T>
-constexpr __host__ __device__ bool isPowerOf2(T v) {
-  return (v && !(v & (v - 1)));
-}
-
-template <bool greater, typename T>
-__forceinline__ __device__ bool is_better_than(T val, T baseline) {
-  return (val > baseline && greater) || (val < baseline && !greater);
-}
-
-template <bool greater, typename T, typename idxT>
-__forceinline__ __device__ bool is_better_than(T val,
-                                               T baseline,
-                                               idxT index,
-                                               idxT baseline_index) {
-  bool res = (val > baseline && greater) || (val < baseline && !greater);
-  if (val == baseline) {
-    res = (index < baseline_index && greater) ||
-          (index < baseline_index && !greater);
-  }
-  return res;
-}
-
-template <typename T, typename idxT>
-int calc_smem_size_for_block_wide(int num_of_warp, int64_t k) {
-  int64_t cache_topk = (sizeof(T) + sizeof(idxT)) * num_of_warp * k;
-  int64_t n = std::max<int>(num_of_warp / 2 * k, num_of_warp * WARP_SIZE);
-  return max(cache_topk,
-             round_up_to_multiple_of<256>(n * sizeof(T)) + n * sizeof(idxT));
-}
-
-template <int size,
-          bool ascending,
-          bool reverse,
-          typename T,
-          typename idxT,
-          bool is_stable>
-struct BitonicMerge {
-  // input should be a bitonic sequence, and sort it to be a monotonic sequence
-  __device__ static void merge(T* __restrict__ val_arr,
-                               idxT* __restrict__ idx_arr) {
-    static_assert(isPowerOf2(size));
-    static_assert(size >= 2 * WARP_SIZE);
-    constexpr int arr_len = size / WARP_SIZE;
-
-    constexpr int stride = arr_len / 2;
-    for (int i = 0; i < stride; ++i) {
-      int const other_i = i + stride;
-      T& val = val_arr[i];
-      T& other_val = val_arr[other_i];
-      bool is_better;
-      if constexpr (is_stable) {
-        is_better = is_better_than<ascending>(
-            val, other_val, idx_arr[i], idx_arr[other_i]);
-      } else {
-        is_better = is_better_than<ascending>(val, other_val);
-      }
-
-      if (is_better) {
-        T tmp = val;
-        val = other_val;
-        other_val = tmp;
-
-        idxT tmp2 = idx_arr[i];
-        idx_arr[i] = idx_arr[other_i];
-        idx_arr[other_i] = tmp2;
-      }
-    }
-
-    BitonicMerge<size / 2, ascending, reverse, T, idxT, is_stable>::merge(
-        val_arr, idx_arr);
-    BitonicMerge<size / 2, ascending, reverse, T, idxT, is_stable>::merge(
-        val_arr + arr_len / 2, idx_arr + arr_len / 2);
-  }
-};
-
-template <int size, bool ascending, typename T, typename idxT, bool is_stable>
-struct BitonicSort {
-  __device__ static void sort(T* __restrict__ val_arr,
-                              idxT* __restrict__ idx_arr) {
-    static_assert(isPowerOf2(size));
-    static_assert(size >= 2 * WARP_SIZE);
-    constexpr int arr_len = size / WARP_SIZE;
-
-    BitonicSort<size / 2, true, T, idxT, is_stable>::sort(val_arr, idx_arr);
-    BitonicSort<size / 2, false, T, idxT, is_stable>::sort(
-        val_arr + arr_len / 2, idx_arr + arr_len / 2);
-    BitonicMerge<size, ascending, ascending, T, idxT, is_stable>::merge(
-        val_arr, idx_arr);
-  }
-};
-
-template <bool ascending, typename T, typename idxT, bool is_stable>
-struct BitonicSort<32, ascending, T, idxT, is_stable> {
-  __device__ static void sort(T* __restrict__ val_arr,
-                              idxT* __restrict__ idx_arr) {
-    int const lane = threadIdx.x % WARP_SIZE;
-
-    // ascending doesn't matter before merging since all we need is a bitonic
-    // sequence
-    for (int stage = 0; stage < 4; ++stage) {
-      for (int stride = (1 << stage); stride > 0; stride /= 2) {
-        bool reverse = (lane >> stage) & 2;
-        bool is_second = lane & stride;
-
-        T other = __shfl_xor_sync(FULL_WARP_MASK, *val_arr, stride);
-        idxT other_idx = __shfl_xor_sync(FULL_WARP_MASK, *idx_arr, stride);
-
-        bool is_better;
-        if constexpr (is_stable) {
-          if constexpr (ascending) {
-            is_better = ((*val_arr > other) ||
-                         ((*val_arr == other) && (*idx_arr < other_idx))) !=
-                        (reverse != is_second);
-          } else {
-            is_better = ((*val_arr > other) ||
-                         ((*val_arr == other) && (*idx_arr > other_idx))) !=
-                        (reverse != is_second);
-          }
-        } else {
-          is_better = (*val_arr != other &&
-                       (*val_arr > other) != (reverse != is_second));
-        }
-        if (is_better) {
-          *val_arr = other;
-          *idx_arr = other_idx;
-        }
-      }
-    }
-
-    BitonicMerge<32, ascending, ascending, T, idxT, is_stable>::merge(val_arr,
-                                                                      idx_arr);
-  }
-};
-
-template <bool ascending,
-          bool reverse,
-          typename T,
-          typename idxT,
-          bool is_stable>
-struct BitonicMerge<32, ascending, reverse, T, idxT, is_stable> {
-  __device__ static void merge(T* __restrict__ val_arr,
-                               idxT* __restrict__ idx_arr) {
-    int const lane = threadIdx.x % WARP_SIZE;
-    for (int stride = WARP_SIZE / 2; stride > 0; stride /= 2) {
-      bool is_second = lane & stride;
-      T& val = *val_arr;
-      T other = __shfl_xor_sync(FULL_WARP_MASK, val, stride);
-      idxT& idx = *idx_arr;
-      idxT other_idx = __shfl_xor_sync(FULL_WARP_MASK, idx, stride);
-
-      bool is_better;
-      if constexpr (is_stable) {
-        if constexpr (ascending) {
-          is_better = ((*val_arr > other) ||
-                       ((*val_arr == other) && (*idx_arr < other_idx))) ==
-                      (reverse != is_second);  // for min
-        } else {
-          is_better = ((*val_arr > other) ||
-                       ((*val_arr == other) && (*idx_arr > other_idx))) ==
-                      (reverse != is_second);  // for max
-        }
-      } else {
-        is_better =
-            (val != other && ((val > other) == (ascending != is_second)));
-      }
-
-      if (is_better) {
-        val = other;
-        idx = other_idx;
-      }
-    }
-  }
-};
-
-template <int capacity, bool greater, typename T, typename idxT, bool is_stable>
-class WarpSort {
- public:
-  __device__ WarpSort(idxT k, T dummy)
-      : lane_(threadIdx.x % WARP_SIZE), k_(k), dummy_(dummy) {
-    static_assert(capacity >= WARP_SIZE && isPowerOf2(capacity));
-
-    for (int i = 0; i < max_arr_len_; ++i) {
-      val_arr_[i] = dummy_;
-      idx_arr_[i] = 0;
-    }
-  }
-
-  // load and merge k sorted values
-  __device__ void load_sorted(T const* __restrict__ in,
-                              idxT const* __restrict__ in_idx,
-                              idxT start) {
-    idxT idx = start + WARP_SIZE - 1 - lane_;
-    for (int i = max_arr_len_ - 1; i >= 0; --i, idx += WARP_SIZE) {
-      if (idx < start + k_) {
-        T t = in[idx];
-        bool is_better;
-        if constexpr (is_stable) {
-          is_better =
-              is_better_than<greater>(t, val_arr_[i], in_idx[idx], idx_arr_[i]);
-        } else {
-          is_better = is_better_than<greater>(t, val_arr_[i]);
-        }
-        if (is_better) {
-          val_arr_[i] = t;
-          idx_arr_[i] = in_idx[idx];
-        }
-      }
-    }
-
-    BitonicMerge<capacity, greater, !greater, T, idxT, is_stable>::merge(
-        val_arr_, idx_arr_);
-  }
-
-  __device__ void dump(T* __restrict__ out, idxT* __restrict__ out_idx) const {
-    for (int i = 0; i < max_arr_len_; ++i) {
-      idxT out_i = i * WARP_SIZE + lane_;
-      if (out_i < k_) {
-        out[out_i] = val_arr_[i];
-        out_idx[out_i] = idx_arr_[i];
-      }
-    }
-  }
-
-  __device__ void dumpIdx(idxT* __restrict__ out_idx) const {
-    for (int i = 0; i < max_arr_len_; ++i) {
-      idxT out_i = i * WARP_SIZE + lane_;
-      if (out_i < k_) {
-        out_idx[out_i] = idx_arr_[i];
-      }
-    }
-  }
-
- protected:
-  static constexpr int max_arr_len_ = capacity / WARP_SIZE;
-
-  T val_arr_[max_arr_len_];
-  idxT idx_arr_[max_arr_len_];
-
-  int const lane_;
-  idxT const k_;
-  T const dummy_;
-
-};  // end class WarpSort
-
-template <int capacity, bool greater, typename T, typename idxT, bool is_stable>
-class WarpSelect : public WarpSort<capacity, greater, T, idxT, is_stable> {
- public:
-  // Default constructor: uses extern __shared__ smem_buf from offset 0.
-  // Used by callers that have no other shared memory before the WarpSelect
-  // staging region.
-  __device__ WarpSelect(idxT k, T dummy)
-      : WarpSort<capacity, greater, T, idxT, is_stable>(k, dummy),
-        k_th_(dummy),
-        k_th_lane_((k - 1) % WARP_SIZE) {
-    extern __shared__ char smem_buf[];
-
-    int const num_of_warp = blockDim.x / WARP_SIZE;
-    int const warp_id = threadIdx.x / WARP_SIZE;
-    val_smem_ = reinterpret_cast<T*>(smem_buf);
-    val_smem_ += warp_id * WARP_SIZE;
-    idx_smem_ = reinterpret_cast<idxT*>(
-        smem_buf +
-        round_up_to_multiple_of<256>(num_of_warp * sizeof(T) * WARP_SIZE));
-    idx_smem_ += warp_id * WARP_SIZE;
-  }
-
-  // Constructor with explicit smem base pointer.
-  // Use this when shared memory before offset `smem_base` must be preserved
-  // (e.g. smem_scores_warp in the fused noaux_tc kernel).
-  __device__ WarpSelect(idxT k, T dummy, char* smem_base)
-      : WarpSort<capacity, greater, T, idxT, is_stable>(k, dummy),
-        k_th_(dummy),
-        k_th_lane_((k - 1) % WARP_SIZE) {
-    int const num_of_warp = blockDim.x / WARP_SIZE;
-    int const warp_id = threadIdx.x / WARP_SIZE;
-    val_smem_ = reinterpret_cast<T*>(smem_base);
-    val_smem_ += warp_id * WARP_SIZE;
-    idx_smem_ = reinterpret_cast<idxT*>(
-        smem_base +
-        round_up_to_multiple_of<256>(num_of_warp * sizeof(T) * WARP_SIZE));
-    idx_smem_ += warp_id * WARP_SIZE;
-  }
-
-  __device__ void add(T const* in, idxT start, idxT end) {
-    idxT const end_for_fullwarp =
-        round_up_to_multiple_of<WARP_SIZE>(end - start) + start;
-    for (idxT i = start + lane_; i < end_for_fullwarp; i += WARP_SIZE) {
-      T val = (i < end) ? in[i] : dummy_;
-      add(val, i);
-    }
-  }
-
-  __device__ void add(T val, idxT idx) {
-    bool do_add;
-    if constexpr (is_stable) {
-      do_add = is_better_than<greater>(val, k_th_, idx, k_th_idx_);
-    } else {
-      do_add = is_better_than<greater>(val, k_th_);
-    }
-
-    uint32_t mask = __ballot_sync(FULL_WARP_MASK, do_add);
-    if (mask == 0) {
-      return;
-    }
-
-    int pos = smem_buf_len_ + __popc(mask & ((0x1u << lane_) - 1));
-    if (do_add && pos < WARP_SIZE) {
-      val_smem_[pos] = val;
-      idx_smem_[pos] = idx;
-      do_add = false;
-    }
-    smem_buf_len_ += __popc(mask);
-    if (smem_buf_len_ >= WARP_SIZE) {
-      __syncwarp();
-      merge_buf_(val_smem_[lane_], idx_smem_[lane_]);
-      smem_buf_len_ -= WARP_SIZE;
-    }
-    if (do_add) {
-      pos -= WARP_SIZE;
-      val_smem_[pos] = val;
-      idx_smem_[pos] = idx;
-    }
-    __syncwarp();
-  }
-
-  __device__ void done() {
-    if (smem_buf_len_) {
-      T val = (lane_ < smem_buf_len_) ? val_smem_[lane_] : dummy_;
-      idxT idx = (lane_ < smem_buf_len_) ? idx_smem_[lane_] : 0;
-      merge_buf_(val, idx);
-    }
-
-    // after done(), smem is used for merging results among warps
-    __syncthreads();
-  }
-
- private:
-  __device__ void set_k_th_() {
-    k_th_ = __shfl_sync(FULL_WARP_MASK, val_arr_[max_arr_len_ - 1], k_th_lane_);
-    if constexpr (is_stable) {
-      k_th_idx_ =
-          __shfl_sync(FULL_WARP_MASK, idx_arr_[max_arr_len_ - 1], k_th_lane_);
-    }
-  }
-
-  __device__ void merge_buf_(T val, idxT idx) {
-    BitonicSort<WARP_SIZE, greater, T, idxT, is_stable>::sort(&val, &idx);
-
-    T& old = val_arr_[max_arr_len_ - 1];
-
-    bool is_better;
-    if constexpr (is_stable) {
-      is_better =
-          is_better_than<greater>(val, old, idx, idx_arr_[max_arr_len_ - 1]);
-    } else {
-      is_better = is_better_than<greater>(val, old);
-    }
-
-    if (is_better) {
-      old = val;
-      idx_arr_[max_arr_len_ - 1] = idx;
-    }
-
-    BitonicMerge<capacity, greater, !greater, T, idxT, is_stable>::merge(
-        val_arr_, idx_arr_);
-
-    set_k_th_();
-  }
-
-  using WarpSort<capacity, greater, T, idxT, is_stable>::max_arr_len_;
-  using WarpSort<capacity, greater, T, idxT, is_stable>::val_arr_;
-  using WarpSort<capacity, greater, T, idxT, is_stable>::idx_arr_;
-  using WarpSort<capacity, greater, T, idxT, is_stable>::lane_;
-  using WarpSort<capacity, greater, T, idxT, is_stable>::k_;
-  using WarpSort<capacity, greater, T, idxT, is_stable>::dummy_;
-
-  T* val_smem_;
-  idxT* idx_smem_;
-  int smem_buf_len_ = 0;
-
-  T k_th_;
-  idxT k_th_idx_;
-  int const k_th_lane_;
-};  // end class WarpSelect
-}  // namespace warp_topk
-
-inline __device__ unsigned int xorwow_moe(unsigned int& state) {
-  state ^= state >> 7;
-  state ^= state << 9;
-  state ^= state >> 13;
-  return state;
-}
-
-template <typename T>
-__device__ void topk_with_k2(T* output,
-                             T const* input,
-                             cg::thread_block_tile<32> const& tile,
-                             int32_t const lane_id,
-                             int const num_experts_per_group  // 32 ~256
-) {
-  // Get the top2 per thread
-  T largest = neg_inf<T>();
-  T second_largest = neg_inf<T>();
-
-  if (num_experts_per_group > WARP_SIZE) {
-    for (int i = lane_id; i < num_experts_per_group; i += WARP_SIZE) {
-      T value = input[i];
-      if (value > largest) {
-        second_largest = largest;
-        largest = value;
-      } else if (value > second_largest) {
-        second_largest = value;
-      }
-    }
-  } else {
-    for (int i = lane_id; i < num_experts_per_group; i += WARP_SIZE) {
-      largest = input[i];
-    }
-  }
-
-  __syncwarp();  // Ensure all threads have valid data before reduction
-  // Get the top2 warpwise
-  T max1 = cg::reduce(tile, largest, cg::greater<T>());
-
-  T max2 = max1;
-  bool equal_to_max1 = (max1 == largest);
-  int count_max1 = __popc(__ballot_sync(FULL_WARP_MASK, equal_to_max1));
-
-  if (count_max1 == 1) {
-    largest = (largest == max1) ? second_largest : largest;
-    max2 = cg::reduce(tile, largest, cg::greater<T>());
-  }
-
-  if (lane_id == 0) {
-    *output = max1 + max2;  // reg <- reg
-  }
-}
-
-template <typename T>
-__device__ void topk_with_k2_kernel(
-    T* output,       // group_scores[n_group]
-    const T* input,  // scores_with_bias[n, num_experts]
-    int64_t const num_tokens,
-    int64_t const n_group,
-    int64_t const num_experts,
-    int64_t const num_experts_per_group) {
-  int32_t warp_id = threadIdx.x / WARP_SIZE;
-  int32_t lane_id = threadIdx.x % WARP_SIZE;
-  int32_t token_id =
-      blockIdx.x * NUM_WARPS_PER_BLOCK + warp_id;  // token_offset
-#if (defined(__CUDA_ARCH__) && (__CUDA_ARCH__ >= 900))
-  asm volatile("griddepcontrol.wait;");
-#endif
-  if (token_id < num_tokens) {
-    cg::thread_block block = cg::this_thread_block();
-    cg::thread_block_tile<32> tile = cg::tiled_partition<32>(block);
-    int32_t input_offset = token_id * num_experts;
-    int32_t output_offset = token_id * n_group;
-    input += input_offset;  // [num_experts]
-#pragma unroll
-    for (int loc_group_idx = 0; loc_group_idx < n_group; loc_group_idx++) {
-      const T* input_prt = input + loc_group_idx * num_experts_per_group;
-      T* output_ptr = output + loc_group_idx;
-      topk_with_k2(output_ptr, input_prt, tile, lane_id, num_experts_per_group);
-    }
-  }
-#if (defined(__CUDA_ARCH__) && (__CUDA_ARCH__ >= 900))
-  asm volatile("griddepcontrol.launch_dependents;");
-#endif
-}
-
-template <typename T, typename IdxT>
-__global__ void group_idx_and_topk_idx_kernel(
-    T* gating_output,
-    T* e_score_correction_bias,
-    T* scores,
-    T* group_scores,
-    T* topk_values,
-    IdxT* topk_indices,
-    int64_t const num_tokens,
-    int64_t const n_group,
-    int64_t const topk_group,
-    int64_t const topk,
-    int64_t const num_experts,
-    int64_t const num_experts_per_group,
-    bool const renormalize,
-    double routed_scaling_factor) {
-  int32_t warp_id = threadIdx.x / WARP_SIZE;
-  int32_t lane_id = threadIdx.x % WARP_SIZE;
-  int32_t case_id =
-      blockIdx.x * NUM_WARPS_PER_BLOCK + warp_id;  // one per token
-
-  extern __shared__ char smem_buf[];
-  // smem layout:
-  //   [0, smem_scores_bytes):  smem_scores[NUM_WARPS_PER_BLOCK * num_experts]
-  //   [smem_scores_bytes, ..): WarpSelect staging, reused as
-  //   s_topk_idx/s_topk_value
-  T* smem_scores_warp =
-      reinterpret_cast<T*>(smem_buf) + warp_id * num_experts;  // [num_experts]
-  char* warpselect_base =
-      smem_buf + NUM_WARPS_PER_BLOCK * num_experts * sizeof(T);
-  int32_t* s_topk_idx =
-      reinterpret_cast<int32_t*>(warpselect_base) + warp_id * topk;
-  T* s_topk_value =
-      reinterpret_cast<T*>(reinterpret_cast<int32_t*>(warpselect_base) +
-                           NUM_WARPS_PER_BLOCK * topk) +
-      warp_id * topk;
-
-  // Register arrays for Phase 0+1 merged loop.
-  // Per-lane top-2 accumulators per group (supports n_group up to 8).
-  constexpr int MAX_N_GROUP = 8;
-  float largest[MAX_N_GROUP];
-  float second_largest[MAX_N_GROUP];
-  // group_scores_reg[g]: warp-reduced top2-sum per group, broadcast to all
-  // lanes. Replaces the global group_scores buffer for Phase 2 and Phase 3.
-  float group_scores_reg[MAX_N_GROUP];
-#pragma unroll
-  for (int g = 0; g < MAX_N_GROUP; g++) {
-    largest[g] = -cuda::std::numeric_limits<float>::infinity();
-    second_largest[g] = -cuda::std::numeric_limits<float>::infinity();
-    group_scores_reg[g] = -cuda::std::numeric_limits<float>::infinity();
-  }
-
-  scores += case_id * num_experts;
-  group_scores += case_id * n_group;
-  topk_values += case_id * topk;
-  topk_indices += case_id * topk;
-  int32_t align_num_experts_per_group =
-      warp_topk::round_up_to_multiple_of<WARP_SIZE>(num_experts_per_group);
-
-  // Phase 0 + Phase 1 merged:
-  // Single pass over [0, num_experts) with stride WARP_SIZE.
-  // Each column i:
-  //   - sigmoid(gating_output[row, i])          → s  (unbiased)
-  //   - s + e_score_correction_bias[i]          → sb (biased)
-  //   - write s to smem_scores_warp[i]          (Phase 3 biased score and Phase
-  //   4 gather use this)
-  //   - update largest/second_largest[g] for g = i / num_experts_per_group
-  if (case_id < num_tokens) {
-    const T* gating_row = gating_output + case_id * num_experts;
-    for (int i = lane_id; i < num_experts; i += WARP_SIZE) {
-      float g_val = cuda_cast<float, T>(gating_row[i]);
-      float s = 1.f / (1.f + __expf(-g_val));
-      float sb = s + cuda_cast<float, T>(e_score_correction_bias[i]);
-
-      smem_scores_warp[i] = cuda_cast<T, float>(s);
-      // topk_with2 收敛到每个lane在某一个group的局部最大值和局部第二最大值
-      int g = i / num_experts_per_group;
-      if (sb > largest[g]) {
-        second_largest[g] = largest[g];
-        largest[g] = sb;
-      } else if (sb > second_largest[g]) {
-        second_largest[g] = sb;
-      }
-    }
-  }
-  __syncwarp();
-
-  // Phase 1 reduction: warp-reduce top-2 per group → store in group_scores_reg.
-  // Result is broadcast to all lanes via __shfl_sync so Phase 2 and Phase 3
-  // can read it directly from registers without touching global group_scores.
-  if (case_id < num_tokens) {
-    cg::thread_block_tile<32> tile_p1 =
-        cg::tiled_partition<32>(cg::this_thread_block());
-    for (int g = 0; g < n_group; g++) {
-      float max1 = cg::reduce(tile_p1, largest[g], cg::greater<float>());
-      float max2 = max1;
-      bool eq = (largest[g] == max1);
-      if (__popc(__ballot_sync(FULL_WARP_MASK, eq)) == 1) {
-        float l = (largest[g] == max1) ? second_largest[g] : largest[g];
-        max2 = cg::reduce(tile_p1, l, cg::greater<float>());
-      }
-      // cg::reduce returns the same value to all lanes in the warp,
-      // so max1+max2 is already consistent across all lanes — store directly.
-      group_scores_reg[g] = max1 + max2;
-    }
-  }
-  // No __syncthreads() needed: group_scores_reg is warp-local register state,
-  // not shared memory. A __syncwarp() suffices to ensure all lanes are past
-  // the reduction before Phase 2 begins.
-  __syncwarp();
-
-  cg::thread_block block = cg::this_thread_block();
-  cg::thread_block_tile<32> tile = cg::tiled_partition<32>(block);
-
-  T value = neg_inf<T>();
-  T topk_group_value = neg_inf<T>();
-  int32_t num_equalto_topkth_group;
-
-#if (defined(__CUDA_ARCH__) && (__CUDA_ARCH__ >= 900))
-  asm volatile("griddepcontrol.wait;");  // I think all prolog can be put before
-                                         // acqbulk because it's ptr arithmetic
-#endif
-
-  if (case_id < num_tokens) {
-    // Phase 2: select topk_group groups using group_scores_reg (register, no
-    // global read). Each lane owns one group slot (lane_id < n_group); lanes >=
-    // n_group stay at -inf.
-    int32_t want_neg_inf_num = WARP_SIZE - n_group + topk_group;
-    if (lane_id < n_group && isfinite(group_scores_reg[lane_id])) {
-      value = cuda_cast<T, float>(group_scores_reg[lane_id]);
-    }
-
-    int neg_inf_num = WARP_SIZE - n_group;
-    int last_neg_inf_num = 0;
-    // Use loop to find the largset top_group
-    while (neg_inf_num < want_neg_inf_num) {
-      __syncwarp();  // Ensure all threads have valid data before reduction
-      topk_group_value = cg::reduce(tile, value, cg::greater<T>());
-      if (value == topk_group_value) {
-        value = neg_inf<T>();
-      }
-      last_neg_inf_num = neg_inf_num;
-
-      neg_inf_num =
-          __popc(__ballot_sync(FULL_WARP_MASK, (value == neg_inf<T>())));
-    }
-    // There is a possible case:
-    // may have many different group holding the same score!
-    // but we only accept some of them!
-    num_equalto_topkth_group = want_neg_inf_num - last_neg_inf_num;
-  }
-  // Only a warp-level sync is needed here since we no longer write to shared
-  // memory in Phase 1; the subsequent WarpSelect constructor does __syncthreads
-  // internally via done(), so this is sufficient.
-  __syncwarp();
-
-  // Pass warpselect_base so WarpSelect staging smem does not overwrite
-  // smem_scores_warp which sits at the start of smem_buf.
-  warp_topk::WarpSelect</*capability*/ WARP_SIZE,
-                        /*greater*/ true,
-                        T,
-                        int32_t,
-                        /* is_stable */ true>
-      queue((int32_t)topk, neg_inf<T>(), warpselect_base);
-
-  int count_equalto_topkth_group = 0;
-  bool if_proceed_next_topk = (topk_group_value != neg_inf<T>());
-  if (case_id < num_tokens && if_proceed_next_topk) {
-    for (int i_group = 0; i_group < n_group; i_group++) {
-      // Use group_scores_reg (register) instead of global group_scores.
-      float gs_i = group_scores_reg[i_group];
-      if ((gs_i > cuda_cast<float, T>(topk_group_value)) ||
-          ((gs_i == cuda_cast<float, T>(topk_group_value)) &&
-           (count_equalto_topkth_group < num_equalto_topkth_group))) {
-        int32_t offset = i_group * num_experts_per_group;
-        for (int32_t i = lane_id; i < align_num_experts_per_group;
-             i += WARP_SIZE) {
-          // Phase 3: compute biased score from smem (unbiased) + global bias.
-          // smem_scores_warp[offset+i] holds sigmoid(gating[offset+i]) written
-          // in Phase 0; adding the bias here recovers the biased score without
-          // needing sbias_reg to be addressable across non-WARP_SIZE-aligned
-          // group boundaries.
-          float sb =
-              (i < num_experts_per_group)
-                  ? (cuda_cast<float, T>(smem_scores_warp[offset + i]) +
-                     cuda_cast<float, T>(e_score_correction_bias[offset + i]))
-                  : -cuda::std::numeric_limits<float>::infinity();
-          T candidates = (i < num_experts_per_group) && isfinite(sb)
-                             ? cuda_cast<T, float>(sb)
-                             : neg_inf<T>();
-          queue.add(candidates, offset + i);
-        }
-        if (gs_i == cuda_cast<float, T>(topk_group_value)) {
-          count_equalto_topkth_group++;
-        }
-      }
-    }
-    queue.done();
-    __syncwarp();
-    // Get the topk_idx
-    queue.dumpIdx(s_topk_idx);
-    __syncwarp();
-  }
-
-  // Phase 4: gather topk_values from smem_scores (unbiased, stored in Phase 0).
-  // No global memory read needed for scores.
-  float topk_sum = 1e-20;
-  if (case_id < num_tokens && if_proceed_next_topk) {
-    for (int i = lane_id;
-         i < warp_topk::round_up_to_multiple_of<WARP_SIZE>(topk);
-         i += WARP_SIZE) {
-      T value = i < topk ? smem_scores_warp[s_topk_idx[i]]
-                         : 0.0f;  // Load the valid value of expert
-      if (i < topk) {
-        s_topk_value[i] = value;
-      }
-      topk_sum +=
-          cg::reduce(tile, cuda_cast<float, T>(value), cg::plus<float>());
-    }
-  }
-
-  __syncthreads();
-  // Note(ZKK): a little trick.
-  if (case_id < num_tokens && if_proceed_next_topk) {
-    for (int i = lane_id; i < num_experts; i += WARP_SIZE) {
-      scores[i] = 0;
-    }
-  }
-  __syncwarp();
-
-  if (case_id < num_tokens) {
-    if (if_proceed_next_topk) {
-      for (int i = lane_id; i < topk; i += WARP_SIZE) {
-        float value;
-        if (renormalize) {
-          value = cuda_cast<float, T>(s_topk_value[i]) / topk_sum *
-                  routed_scaling_factor;
-        } else {
-          value = cuda_cast<float, T>(s_topk_value[i]) * routed_scaling_factor;
-        }
-        scores[s_topk_idx[i]] = value;
-        topk_indices[i] = s_topk_idx[i];
-        topk_values[i] = cuda_cast<T, float>(value);
-      }
-    } else {
-      for (int i = lane_id; i < topk; i += WARP_SIZE) {
-        topk_indices[i] = i;
-        topk_values[i] = cuda_cast<T, float>(1.0f / topk);
-      }
-    }
-    // Note: when if_proceed_next_topk==false, choose the first 8 experts as the
-    // default result.
-  }
-#if (defined(__CUDA_ARCH__) && (__CUDA_ARCH__ >= 900))
-  asm volatile("griddepcontrol.launch_dependents;");
-#endif
-}
-
-template <typename T, typename IdxT>
-__global__ void group_idx_and_topk_idx_redundant_kernel(
-    T* scores,
-    T const* group_scores,
-    T* topk_values,
-    IdxT* topk_indices,
-    T* scores_with_bias,
-    int32_t* expert_id_to_ep_rank_array,
-    int32_t* expert_in_rank_num_list,
-    int32_t* tokens_per_expert_stats_list,
-    int64_t const num_tokens,
-    int64_t const n_group,
-    int64_t const topk_group,
-    int64_t const topk,
-    bool const renormalize,
-    int64_t const num_experts,
-    int64_t const num_experts_per_group,
-    double routed_scaling_factor,
-    int64_t const redundant_ep_rank_num_plus_one) {
-  int32_t warp_id = threadIdx.x / WARP_SIZE;
-  int32_t lane_id = threadIdx.x % WARP_SIZE;
-  int32_t case_id =
-      blockIdx.x * NUM_WARPS_PER_BLOCK + warp_id;  // one per token
-  unsigned int state = case_id;
-  scores_with_bias += case_id * num_experts;
-  scores += case_id * num_experts;
-  group_scores += case_id * n_group;
-  topk_values += case_id * topk;
-  topk_indices += case_id * topk;
-  int32_t align_num_experts_per_group =
-      warp_topk::round_up_to_multiple_of<WARP_SIZE>(num_experts_per_group);
-
-  cg::thread_block block = cg::this_thread_block();
-  cg::thread_block_tile<32> tile = cg::tiled_partition<32>(block);
-
-  extern __shared__ char smem_buf[];  // NOTE: reuse the shared memory here to
-                                      // store the target topk idx
-  int32_t* s_topk_idx = reinterpret_cast<int32_t*>(smem_buf);
-  T* s_topk_value =
-      reinterpret_cast<T*>(s_topk_idx + NUM_WARPS_PER_BLOCK * topk) +
-      warp_id * topk;
-  s_topk_idx += warp_id * topk;
-
-  T value = neg_inf<T>();
-  T topk_group_value = neg_inf<T>();
-  int32_t num_equalto_topkth_group;
-
-#if (defined(__CUDA_ARCH__) && (__CUDA_ARCH__ >= 900))
-  asm volatile("griddepcontrol.wait;");  // I think all prolog can be put before
-                                         // acqbulk because it's ptr arithmetic
-#endif
-
-  if (case_id < num_tokens) {
-    // calculate group_idx
-    int32_t target_num_min = WARP_SIZE - n_group + topk_group;
-    if (lane_id < n_group &&
-        (isfinite(cuda_cast<float, T>(
-            group_scores[lane_id]))))  // The check is necessary to avoid
-                                       // abnormal input
-    {
-      value = group_scores[lane_id];
-    }
-
-    int count_equal_to_top_value = WARP_SIZE - n_group;
-    int pre_count_equal_to_top_value = 0;
-    // Use loop to find the largset top_group
-    while (count_equal_to_top_value < target_num_min) {
-      __syncwarp();  // Ensure all threads have valid data before reduction
-      topk_group_value = cg::reduce(tile, value, cg::greater<T>());
-      if (value == topk_group_value) {
-        value = neg_inf<T>();
-      }
-      pre_count_equal_to_top_value = count_equal_to_top_value;
-      count_equal_to_top_value =
-          __popc(__ballot_sync(FULL_WARP_MASK, (value == neg_inf<T>())));
-    }
-    num_equalto_topkth_group = target_num_min - pre_count_equal_to_top_value;
-  }
-  __syncthreads();
-
-  warp_topk::WarpSelect</*capability*/ WARP_SIZE,
-                        /*greater*/ true,
-                        T,
-                        int32_t,
-                        /* is_stable */ true>
-      queue((int32_t)topk, neg_inf<T>());
-
-  int count_equalto_topkth_group = 0;
-  bool if_proceed_next_topk = (topk_group_value != neg_inf<T>());
-  if (case_id < num_tokens && if_proceed_next_topk) {
-    for (int i_group = 0; i_group < n_group; i_group++) {
-      if ((group_scores[i_group] > topk_group_value) ||
-          ((group_scores[i_group] == topk_group_value) &&
-           (count_equalto_topkth_group < num_equalto_topkth_group))) {
-        int32_t offset = i_group * num_experts_per_group;
-        for (int32_t i = lane_id; i < align_num_experts_per_group;
-             i += WARP_SIZE) {
-          T candidates =
-              (i < num_experts_per_group) && isfinite(cuda_cast<float, T>(
-                                                 scores_with_bias[offset + i]))
-                  ? scores_with_bias[offset + i]
-                  : neg_inf<T>();
-          queue.add(candidates, offset + i);
-        }
-        if (group_scores[i_group] == topk_group_value) {
-          count_equalto_topkth_group++;
-        }
-      }
-    }
-    queue.done();
-    __syncwarp();
-    // Get the topk_idx
-    queue.dumpIdx(s_topk_idx);
-    __syncwarp();
-  }
-
-  // Load the valid score value
-  // Calculate the summation
-  float topk_sum = 1e-20;
-  if (case_id < num_tokens && if_proceed_next_topk) {
-    for (int i = lane_id;
-         i < warp_topk::round_up_to_multiple_of<WARP_SIZE>(topk);
-         i += WARP_SIZE) {
-      T value = i < topk ? scores[s_topk_idx[i]]
-                         : 0.0f;  // Load the valid value of expert
-      if (i < topk) {
-        s_topk_value[i] = value;
-      }
-      topk_sum +=
-          cg::reduce(tile, cuda_cast<float, T>(value), cg::plus<float>());
-    }
-  }
-
-  __syncthreads();
-  // Note(ZKK): a little trick.
-  if (case_id < num_tokens && if_proceed_next_topk) {
-    for (int i = lane_id; i < num_experts; i += WARP_SIZE) {
-      scores[i] = 0;
-    }
-  }
-  __syncwarp();
-
-  if (case_id < num_tokens) {
-    if (if_proceed_next_topk) {
-      for (int i = lane_id; i < topk; i += WARP_SIZE) {
-        float value;
-        if (renormalize) {
-          value = cuda_cast<float, T>(s_topk_value[i]) / topk_sum *
-                  routed_scaling_factor;
-        } else {
-          value = cuda_cast<float, T>(s_topk_value[i]) * routed_scaling_factor;
-        }
-        scores[s_topk_idx[i]] = value;
-
-        int expert_topk = s_topk_idx[i];
-        int len = expert_in_rank_num_list[expert_topk];
-        int select = (int)xorwow_moe(state) % len;
-        // int select = 0;
-        int selected_rank =
-            expert_id_to_ep_rank_array[expert_topk *
-                                           redundant_ep_rank_num_plus_one +
-                                       select];
-        atomicAdd(&tokens_per_expert_stats_list[expert_topk], 1);
-        topk_indices[i] = (IdxT)selected_rank;
-        topk_values[i] = cuda_cast<T, float>(value);
-      }
-    } else {
-      for (int i = lane_id; i < topk; i += WARP_SIZE) {
-        int expert_topk = i;
-        int len = expert_in_rank_num_list[expert_topk];
-        int select = (int)xorwow_moe(state) % len;
-        // int select = 0;
-        int selected_rank =
-            expert_id_to_ep_rank_array[expert_topk *
-                                           redundant_ep_rank_num_plus_one +
-                                       select];
-        atomicAdd(&tokens_per_expert_stats_list[expert_topk], 1);
-        topk_indices[i] = (IdxT)selected_rank;
-        topk_values[i] = cuda_cast<T, float>(1.0f / topk);
-      }
-    }
-    // Note: when if_proceed_next_topk==false, choose the first 8 experts as the
-    // default result.
-  }
-#if (defined(__CUDA_ARCH__) && (__CUDA_ARCH__ >= 900))
-  asm volatile("griddepcontrol.launch_dependents;");
-#endif
-}
-
-template <typename T, typename IdxT>
-void invokeNoAuxTc(T* gating_output,
-                   T* bias,
-                   T* scores,
-                   T* group_scores,
-                   T* topk_values,
-                   IdxT* topk_indices,
-                   int64_t const num_tokens,
-                   int64_t const num_experts,
-                   int64_t const n_group,
-                   int64_t const topk_group,
-                   int64_t const topk,
-                   bool const renormalize,
-                   double const routed_scaling_factor,
-                   cudaStream_t const stream) {
-  //   int64_t num_cases = num_tokens * n_group;
-  //   int64_t topk_with_k2_num_blocks = (num_cases - 1) / NUM_WARPS_PER_BLOCK +
-  //   1;
-
-  // #ifdef PADDLE_WITH_CUSTOM_DEVICE_METAX_GPU
-  //   topk_with_k2_kernel<T><<<topk_with_k2_num_blocks, BLOCK_SIZE, 0,
-  //   stream>>>(
-  //       group_scores,
-  //       scores_with_bias,
-  //       num_cases,
-  //       n_group,
-  //       num_experts / n_group);
-  // #else
-  //   auto* kernel_instance1 = &topk_with_k2_kernel<T>;
-  //   cudaLaunchConfig_t config;
-  //   config.gridDim = topk_with_k2_num_blocks;
-  //   config.blockDim = BLOCK_SIZE;
-  //   config.dynamicSmemBytes = 0;
-  //   config.stream = stream;
-  //   cudaLaunchAttribute attrs[1];
-  //   attrs[0].id = cudaLaunchAttributeProgrammaticStreamSerialization;
-  //   attrs[0].val.programmaticStreamSerializationAllowed = false;
-  //   config.numAttrs = 1;
-  //   config.attrs = attrs;
-  //   cudaLaunchKernelEx(&config,
-  //                      kernel_instance1,
-  //                      group_scores,
-  //                      scores_with_bias,
-  //                      num_cases,
-  //                      n_group,
-  //                      num_experts / n_group);
-  // #endif
-  int64_t topk_with_k_group_num_blocks =
-      (num_tokens - 1) / NUM_WARPS_PER_BLOCK + 1;
-  size_t smem_score_bytes = NUM_WARPS_PER_BLOCK * num_experts * sizeof(T);
-  size_t dynamic_smem_in_bytes =
-      smem_score_bytes + warp_topk::calc_smem_size_for_block_wide<T, int32_t>(
-                             NUM_WARPS_PER_BLOCK, topk);
-
-#ifdef PADDLE_WITH_CUSTOM_DEVICE_METAX_GPU
-  group_idx_and_topk_idx_kernel<T, IdxT><<<topk_with_k_group_num_blocks,
-                                           BLOCK_SIZE,
-                                           dynamic_smem_in_bytes,
-                                           stream>>>(gating_output,
-                                                     bias,
-                                                     scores,
-                                                     group_scores,
-                                                     topk_values,
-                                                     topk_indices,
-                                                     num_tokens,
-                                                     n_group,
-                                                     topk_group,
-                                                     topk,
-                                                     num_experts,
-                                                     num_experts / n_group,
-                                                     renormalize,
-                                                     routed_scaling_factor);
-#else
-  cudaLaunchConfig_t config;
-  cudaLaunchAttribute attrs[1];
-  auto* kernel_instance2 = &group_idx_and_topk_idx_kernel<T, IdxT>;
-  config.gridDim = topk_with_k_group_num_blocks;
-  config.blockDim = BLOCK_SIZE;
-  config.dynamicSmemBytes = dynamic_smem_in_bytes;
-  config.stream = stream;
-  attrs[0].id = cudaLaunchAttributeProgrammaticStreamSerialization;
-  attrs[0].val.programmaticStreamSerializationAllowed = false;
-  config.numAttrs = 1;
-  config.attrs = attrs;
-  cudaLaunchKernelEx(&config,
-                     kernel_instance2,
-                     gating_output,
-                     bias,
-                     scores,
-                     group_scores,
-                     topk_values,
-                     topk_indices,
-                     num_tokens,
-                     n_group,
-                     topk_group,
-                     topk,
-                     num_experts,
-                     num_experts / n_group,
-                     renormalize,
-                     routed_scaling_factor);
-#endif
-}
-
-template <typename T, typename IdxT>
-void invokeNoAuxTcRedundant(T* scores,
-                            T* group_scores,
-                            T* topk_values,
-                            IdxT* topk_indices,
-                            T* scores_with_bias,
-                            int32_t* expert_id_to_ep_rank_array,
-                            int32_t* expert_in_rank_num_list,
-                            int32_t* tokens_per_expert_stats_list,
-                            int64_t const num_tokens,
-                            int64_t const num_experts,
-                            int64_t const n_group,
-                            int64_t const topk_group,
-                            int64_t const topk,
-                            bool const renormalize,
-                            double const routed_scaling_factor,
-                            int64_t const redundant_ep_rank_num_plus_one,
-                            cudaStream_t const stream) {
-  int64_t num_cases = num_tokens * n_group;
-  int64_t topk_with_k2_num_blocks = (num_cases - 1) / NUM_WARPS_PER_BLOCK + 1;
-
-  // topk_with_k2_kernel<T><<<topk_with_k2_num_blocks, BLOCK_SIZE, 0, stream>>>(
-  //     group_scores,
-  //     scores_with_bias,
-  //     num_cases,
-  //     n_group,
-  //     num_experts / n_group);
-
-  int64_t topk_with_k_group_num_blocks =
-      (num_tokens - 1) / NUM_WARPS_PER_BLOCK + 1;
-  size_t dynamic_smem_in_bytes =
-      warp_topk::calc_smem_size_for_block_wide<T, int32_t>(NUM_WARPS_PER_BLOCK,
-                                                           topk);
-
-  group_idx_and_topk_idx_redundant_kernel<T>
-      <<<topk_with_k_group_num_blocks,
-         BLOCK_SIZE,
-         dynamic_smem_in_bytes,
-         stream>>>(scores,
-                   group_scores,
-                   topk_values,
-                   topk_indices,
-                   scores_with_bias,
-                   expert_id_to_ep_rank_array,
-                   expert_in_rank_num_list,
-                   tokens_per_expert_stats_list,
-                   num_tokens,
-                   n_group,
-                   topk_group,
-                   topk,
-                   renormalize,
-                   num_experts,
-                   num_experts / n_group,
-                   routed_scaling_factor,
-                   redundant_ep_rank_num_plus_one);
-}
-
-#define INSTANTIATE_NOAUX_TC(T, IdxT)                                      \
-  template void invokeNoAuxTc<T, IdxT>(T * gating_output,                  \
-                                       T * bias,                           \
-                                       T * scores,                         \
-                                       T * group_scores,                   \
-                                       T * topk_values,                    \
-                                       IdxT * topk_indices,                \
-                                       int64_t const num_tokens,           \
-                                       int64_t const num_experts,          \
-                                       int64_t const n_group,              \
-                                       int64_t const topk_group,           \
-                                       int64_t const topk,                 \
-                                       bool const renormalize,             \
-                                       double const routed_scaling_factor, \
-                                       cudaStream_t const stream);
-
-INSTANTIATE_NOAUX_TC(float, int32_t);
-
-#define INSTANTIATE_NOAUX_TC_Redundant(T, IdxT)     \
-  template void invokeNoAuxTcRedundant<T, IdxT>(    \
-      T * scores,                                   \
-      T * group_scores,                             \
-      T * topk_values,                              \
-      IdxT * topk_indices,                          \
-      T * scores_with_bias,                         \
-      int32_t* expert_id_to_ep_rank_array,          \
-      int32_t* expert_in_rank_num_list,             \
-      int32_t* tokens_per_expert_stats_list,        \
-      int64_t const num_tokens,                     \
-      int64_t const num_experts,                    \
-      int64_t const n_group,                        \
-      int64_t const topk_group,                     \
-      int64_t const topk,                           \
-      bool const renormalize,                       \
-      double const routed_scaling_factor,           \
-      int64_t const redundant_ep_rank_num_plus_one, \
-      cudaStream_t const stream);
-
-INSTANTIATE_NOAUX_TC_Redundant(float, int32_t);
